crypto: msm: remove wakeup lock in qcrypto driver
When log server still runs and try to enable CE clock just before sleep,
wakeup lock will prevent the device sleep. And the bam/crypto is a not a
wakeupable device. As we have a state machine in crypto driver, and it
is able to control suspend behavior well, we do not need to use wakeup
lock in qcrypto driver.

Signed-off-by: Zhen Kong <zkong@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>